Create main.go
Create a console application "Calculator". The application must read from the console the strings entered by the user, numbers, arithmetic operations performed between them, and output the result of their execution to the console.
The calculator can perform addition, subtraction, multiplication and division operations with two numbers: a + b, a - b, a * b, a / b. Data is transmitted in one line (see the example below). Solutions in which every number and arithmetic operation is passed from a new line are considered incorrect.
The calculator can work with both Arabic (1,2,3,4,5...) and Roman (I,II,III,IV,V...) numbers.
The calculator must accept numbers from 1 to 10 inclusive, no more. At the output, the numbers are not limited in magnitude and can be any.
The calculator can only work with integers.
The calculator can only work with Arabic or Roman numerals at the same time, when the user enters a line like 3 + II, the calculator must indicate an error and stop working.
When entering Roman numbers, the answer should be output in Roman numerals, respectively, when entering Arabic, the answer is expected in Arabic.
When the user enters the wrong numbers, the application outputs an error to the terminal and shuts down.
When the user enters a string that does not correspond to one of the arithmetic operations described above, the application outputs an error and shuts down.
The result of the division operation is an integer, the remainder is discarded.
The result of the calculator with Arabic numbers can be negative numbers and zero. The result of the calculator with Roman numerals can only be positive numbers, if the result is less than one, the program must indicate an exception.